% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_underreport_estimates.R
\name{generate_underreporting_estimates}
\alias{generate_underreporting_estimates}
\title{Generate the Under Reporting Estimates}
\usage{
generate_underreporting_estimates(
  dat,
  mean_low = 8.7,
  median_low = 6.7,
  mean_mid = 13,
  median_mid = 9.1,
  mean_high = 20.9,
  median_high = 13.7,
  ...
)
}
\arguments{
\item{dat}{the case data}

\item{mean_low}{mean delay days}

\item{median_low}{median delay days}

\item{mean_mid}{mean delay days}

\item{median_mid}{median delay days}

\item{mean_high}{mean delay days}

\item{median_high}{median delay days}

\item{...}{additional arguments to pass to \code{under_reporting_estimates}}
}
\description{
Generates the under reporting estimates based off of reporting delays
and the assumed case fatality rates.
}
