% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalisation_to_death_truncated.R
\name{hospitalisation_to_death_truncated}
\alias{hospitalisation_to_death_truncated}
\title{Hospitalization to Death Truncated}
\usage{
hospitalisation_to_death_truncated(x, mu, sigma)
}
\description{
Hospitalization to Death Truncated
}
