\name{plotHospitals}
\Rdversion{1.1}
\alias{plotHospitals}
\title{
Plot simulated hospital loads
}
\description{
Plots expected hospital loads and prediction intervals
}
\usage{
plotHospitals(x)
}
\arguments{
  \item{x}{simulated hospital loads
}
}
\seealso{
\code{link{simHospitals}}
}
