% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_state_cases.R
\name{pull_state_cases}
\alias{pull_state_cases}
\title{Retrieve State Data}
\usage{
pull_state_cases(state_of_interest = "North Carolina", min_reported_deaths = 5)
}
\arguments{
\item{state_of_interest}{the state desired}
}
\description{
This function pull state data and formats for case reporting
}
