% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cfr.R
\name{scale_cfr}
\alias{scale_cfr}
\title{Function to work out correction CFR}
\usage{
scale_cfr(data_1_in, delay_fun)
}
\arguments{
\item{data_1_in}{the input data}

\item{delay_fun}{the delay function to use in reporting}
}
\description{
Function to work out correction CFR
}
