% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sigma_transform}
\alias{sigma_transform}
\title{Transform for Sigma}
\usage{
sigma_transform(zMean, mu)
}
\arguments{
\item{zMean}{the median}

\item{mu}{the median}
}
\description{
Transform for Sigma
}
