\name{simHospitals}
\Rdversion{1.1}
\alias{simHospitals}

\title{
Generating Hopstial Times
}
\description{
\code{simHospitals} is a function used to simulate the hospital times. 
}
\usage{
simHospitals(paramSim, Ndays = 100, NinfectionsPerDay = 10)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{paramSim}{the sample of parameters simulated from \code{mcmcPandemic} function.}
  \item{Ndays}{number of research days.}
  \item{NinfectionsPerDay}{number of infections per day.}
}
\details{

}
\value{

}
\references{

}
\author{
Patrick Brown
}
\note{

}



\seealso{
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ }
\keyword{ }% __ONLY ONE__ keyword per line
