% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/under_reporting_estimates.R
\name{under_reporting_estimates}
\alias{under_reporting_estimates}
\title{Estimate Under Reporting}
\usage{
under_reporting_estimates(
  data,
  delay_fun,
  cCFRBaseline = 1.4,
  cCFREstimateRange = c(1.2, 1.7)
)
}
\arguments{
\item{data}{the case data to be analyzed}

\item{delay_fun}{the delay function to use to adjust for reporting delay}

\item{cCFRBaseline}{case fatality rate base line}

\item{cCFREstimateRange}{case fatality estimate range}
}
\description{
This function performs an exact binomial test in order
to estimate the degree to which underestimation of cases is occuring
}
