% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalisation_to_death_truncated.R
\name{hospitalisation_to_death_truncated}
\alias{hospitalisation_to_death_truncated}
\title{Hospitalization to Death Truncated}
\usage{
hospitalisation_to_death_truncated(x, mu, sigma)
}
\arguments{
\item{x}{the input value}

\item{mu}{the mean of the lognormal function}

\item{sigma}{the standard deviation of the lognomal distribution}
}
\description{
Hospitalization to Death Truncated
}
